I3Direction and I3Position now store floats instead of doubles as their internal and serialized state. Also cleaned up the classes a bit and moved simple functions from cxx to header
git-svn-id: http://code.icecube.wisc.edu/svn/projects/dataclasses/branches/slimclasses@84124 16731396-06f5-0310-8873-f7f720988828

+++ private/pybindings/I3Direction.cxx
@@ -30,11 +30,11 @@ void register_I3Direction()
 {
 
   void (I3Direction::* oneary)(const I3Direction&) = &I3Direction::SetDirection;
-  void (I3Direction::* twoary)(double, double) = &I3Direction::SetDirection;
+  void (I3Direction::* twoary)(float, float) = &I3Direction::SetDirection;
   void (I3Direction::* threeary)(double, double, double) = &I3Direction::SetDirection;
 
   class_<I3Direction, bases<I3FrameObject>, boost::shared_ptr<I3Direction> >("I3Direction")
-    .def(init<double,double>())
+    .def(init<float,float>())
     .def(init<double,double,double>())
     .def("set_direction", oneary)
     .def("set_direction", twoary)
